Phase 4

+++ Services/PostService.cs
@@ -36,12 +36,17 @@ namespace Blogger.Services
             fetchedPost.Title = post.Title;
             fetchedPost.Body = post.Body;
             fetchedPost.CreatedOn = post.CreatedOn;
+            fetchedPost.LastUpdateOn = DateTime.Now;
             return await PostRepo.Update(fetchedPost);
         }
 
         public async Task<Post> Delete(Guid id)
         {
             var post = await PostRepo.Get(id);
+            if(post == null)
+            {
+                return null;
+            }
             return await PostRepo.Delete(post);
         }
